Comments and Brom
Brom is a Heavy Infantryman
Brom thought this would be a quiet post where he could put his feet up
Alas, Brom was wrong

Brom is an infantryman that remains on the north bridge of the fortress,
serving as a cushion between the invading orcs
and those orcs stomping all over the fort.
The other option would be Making The Map Longer, but this gives slight warning
Well. It could.